platforms/aws-lambda/python/requirements.txt
python-dotenvx
